+++ src/RcsbSearch/SearchRequest.ts
@@ -8,8 +8,16 @@ import {LocalStorageTools as LST} from "../RcsbLocalStorage/LocalStorageTools";
 export class SearchRequest {
     private readonly uri: string;
     private readonly fetch:(input:RequestInfo,init?:RequestInit)=>Promise<Response>;
-    constructor(uri?:string, externalFetch?:(input:RequestInfo,init?:RequestInit)=>Promise<Response>) {
+    private readonly requestConfig: RequestInit;
+    constructor(uri?:string, externalFetch?:(input:RequestInfo,init?:RequestInit)=>Promise<Response>, requestConfig?: RequestInit) {
         this.uri = uri ?? serverSearch.uri;
+        this.requestConfig =  {
+            headers:{
+                'Accept': 'application/json',
+                'Content-Type': 'application/json'
+            },
+            ...(requestConfig ?? {})
+        };
         if(typeof externalFetch === "function")
             this.fetch = externalFetch;
         else
@@ -17,18 +25,19 @@ export class SearchRequest {
         if(!this.fetch)
             throw "ERROR: fetch function was not provided"
     }
-    public async request(query: SearchQuery): Promise<QueryResult|null>{
+    public async request(query: SearchQuery, headers: HeadersInit = {}): Promise<QueryResult|null>{
         const localObj: QueryResult | null = LST.getItem<SearchQuery,QueryResult|null>(query);
         if(localObj)
             return localObj;
         //TODO this fetch call is needed to avoid the error [TypeError: 'fetch' called on an object that does not implement interface Window.]
         const response: Response = await (globalThis.window?.fetch ?? this.fetch)(this.uri, {
+            ...this.requestConfig,
             method: 'POST',
+            body: JSON.stringify(query),
             headers: {
-                'Accept': 'application/json',
-                'Content-Type': 'application/json'
-            },
-            body: JSON.stringify(query)
+                ...this.requestConfig.headers,
+                ...headers
+            }
         });
         try {
             const queryResult: QueryResult = await response.json() as QueryResult;